Style: 상세화면 코드 스타일 수정
- private 설정을 위해 constarint 설정 메소드 같은 파일내로 옮김
- 구분을 위해 마크다운 추가
