+++ iOS/SecondHand/SecondHand/Detail/UI/DetailContentView.swift
@@ -42,13 +42,16 @@ class DetailContentView: UIView {
             새 제품으로 보존된 제품으로 전용박스까지 보내드립니다.\n
             사이즈는 235 입니다.\n fdjkgnsdfjkngsdfjknsgfkjnadfsgjnk
         """
-        productInfo.configure(name: productName,
-                              annotation: annotation,
-                              description: description)
-        
-        communicationInfo.configure(chatCount: 0,
-                                    favoriteCount: 0,
-                                    viewsCount: 1)
+        productInfo.configure(
+            name: productName,
+            annotation: annotation,
+            description: description
+        )
+        communicationInfo.configure(
+            chatCount: 0,
+            favoriteCount: 0,
+            viewsCount: 1
+        )
     }
     
     private func setUI() {
@@ -67,11 +70,90 @@ class DetailContentView: UIView {
         button.layer.borderColor = UIColor.gray.cgColor
         button.layer.borderWidth = 0.2
         
-        let onReservation = UIAction(title: "예약중", handler: { _ in return })
-        let onSold = UIAction(title: "판매완료", handler: { _ in return })
+        let onReservation = UIAction(
+            title: "예약중",
+            handler: { _ in return })
+        let onSold = UIAction(
+            title: "판매완료",
+            handler: { _ in return }
+        )
         
-        button.menu = UIMenu(options: .displayInline,
-                             children: [onReservation, onSold])
+        button.menu = UIMenu(
+            options: .displayInline,
+            children: [onReservation, onSold])
         return button
     }
 }
+
+// MARK: - Constraint 설정
+extension DetailContentView {
+    private func addSubviews() {
+        let subViews = [
+            productImageView,
+            pageControl,
+            sellerInfo,
+            statusButton,
+            productInfo
+        ]
+        
+        subViews.forEach {
+            self.addSubview($0)
+            $0.translatesAutoresizingMaskIntoConstraints = false
+        }
+    }
+    
+    private func layoutConstraint() {
+        addSubviews()
+        layoutProductImageView()
+        layoutPageControl()
+        layoutSellerInfo()
+        layoutStatusButton()
+        layoutProductInfo()
+    }
+    
+    private func layoutProductImageView() {
+        let heightRatio: CGFloat = 5 / 4
+        NSLayoutConstraint.activate([
+            productImageView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
+            productImageView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+            productImageView.topAnchor.constraint(equalTo: self.topAnchor),
+            productImageView.heightAnchor.constraint(equalTo: productImageView.widthAnchor,
+                                                     multiplier: heightRatio),
+        ])
+    }
+    
+    private func layoutPageControl() {
+        NSLayoutConstraint.activate([
+            pageControl.leadingAnchor.constraint(equalTo: self.productImageView.leadingAnchor),
+            pageControl.trailingAnchor.constraint(equalTo: self.productImageView.trailingAnchor),
+            pageControl.bottomAnchor.constraint(equalTo: productImageView.bottomAnchor),
+            pageControl.heightAnchor.constraint(equalToConstant: 20),
+        ])
+    }
+    
+    private func layoutSellerInfo() {
+        NSLayoutConstraint.activate([
+            sellerInfo.leadingAnchor.constraint(equalTo: productImageView.leadingAnchor, constant: 16),
+            sellerInfo.trailingAnchor.constraint(equalTo: productImageView.trailingAnchor, constant: -16),
+            sellerInfo.topAnchor.constraint(equalTo: self.pageControl.bottomAnchor, constant: 16),
+            sellerInfo.heightAnchor.constraint(equalToConstant: 60)
+        ])
+    }
+    
+    private func layoutStatusButton() {
+        NSLayoutConstraint.activate([
+            statusButton.leadingAnchor.constraint(equalTo: sellerInfo.leadingAnchor),
+            statusButton.topAnchor.constraint(equalTo: sellerInfo.bottomAnchor, constant: 16)
+        ])
+    }
+    
+    private func layoutProductInfo() {
+        NSLayoutConstraint.activate([
+            productInfo.leadingAnchor.constraint(equalTo: sellerInfo.leadingAnchor),
+            productInfo.trailingAnchor.constraint(lessThanOrEqualTo: sellerInfo.trailingAnchor),
+            productInfo.topAnchor.constraint(equalTo: statusButton.bottomAnchor, constant: 16),
+            productInfo.bottomAnchor.constraint(equalTo: self.bottomAnchor)
+        ])
+    }
+}
+
